+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## 0.1.3 - 2018-10-31
+
+### Added
+
+ - Ability to mount exporter to Rails application routes as standalone Rack app. @Envek
+
+   ```ruby
+   Rails.application.routes.draw do
+     mount Yabeda::Prometheus::Exporter => "/metrics"
+   end
+   ```
+
 ## 0.1.2 - 2018-10-17
 
 ### Changed